+++ client/src/components/header/index.tsx
@@ -1,7 +1,7 @@
 import { HStack, Flex, Heading, Text, Button, Spacer } from '@chakra-ui/react';
 import { useRouter } from 'next/router';
 import { useState } from 'react';
-import Fonts from '../../public/fonts/GangwonEduPower';
+import Fonts from '../../../public/fonts/GangwonEduPower';
 
 export function Header() {
   const [isLogin, setIsLogin] = useState(false);
