+++ client/src/components/signup/Form/index.tsx
@@ -1,8 +1,8 @@
 import { Flex, FormLabel, Text } from '@chakra-ui/react';
 import { ChangeEvent, useState } from 'react';
-import { EMAIL_BUTTON_THEME } from '../../constants/theme';
-import InputForm from '../util/InputForm';
-import SubmitButton from '../util/SubmitButton';
+import { EMAIL_BUTTON_THEME } from '../../../styles/theme';
+import InputForm from '../../common/InputForm';
+import SubmitButton from '../../common/SubmitButton';
 
 export default function SignupForm() {
   const [signupInputs, setSignupInputs] = useState({
