+++ client/src/components/signup/Tos/index.tsx
