+++ client/src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import type { AppProps } from 'next/app';
 import { ChakraProvider, extendTheme, Spacer } from '@chakra-ui/react';
-import { Header } from '../components/Header';
-import { Footer } from '../components/Footer';
+import { Header } from '../components/header';
+import { Footer } from '../components/footer';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
