+++ client/src/pages/signup.tsx
@@ -1,6 +1,6 @@
-import SignupTos from '../components/signup/SignupTos';
-import SignupLayout from '../components/signup/SignupLayout';
-import SignupForm from '../components/signup/SignupForm';
+import SignupTos from '../components/signup/Tos';
+import SignupLayout from '../components/signup/Layout';
+import SignupForm from '../components/signup/Form';
 
 export default function SignupPage() {
   return (
